refactor: Remove warnings

+++ coreClasses/scene/MapGenerator.cpp
@@ -875,12 +875,10 @@ void MapGenerator::verificationMap (std::ifstream& input, const Map& map) {
                 unsigned int typeOfBlock = 0;
 
                 if (block) {
-                    typeOfBlock = static_cast <unsigned int>(
-                        map.getType({ x, y, z }));
-                    output << static_cast <const unsigned int>(typeOfBlock);
+                    typeOfBlock = static_cast <unsigned int>( map.getType({ x, y, z }));
+                    output << typeOfBlock;
                 } else {
-                    output << static_cast <const unsigned int>
-                        (Map::BlockTypes::None);
+                    output << static_cast <unsigned int> (Map::BlockTypes::None);
                 }
                 if (typeOfBlock == 4 || typeOfBlock == 6) {
                     output << " ";
@@ -936,12 +934,11 @@ void MapGenerator::verificationMap (std::ifstream& input, const Map& map) {
                         unsigned int typeOfBlock =
                             static_cast <unsigned int>(
                                 map.getType({ x, y, z }));
-                        output << static_cast <const unsigned int>(typeOfBlock);
+                        output << typeOfBlock;
                     }
                     if (x != map.width() - 1) output << " ";
                 } else {
-                    output << static_cast <const unsigned int>
-                        (Map::BlockTypes::None);
+                    output << static_cast <unsigned int> (Map::BlockTypes::None);
                     if (x != map.width() - 1) output << " ";
                 }
             }
@@ -962,8 +959,7 @@ void MapGenerator::verificationMap (std::ifstream& input, const Map& map) {
                     currentInfo >= map.getEnemiesInfo().size() ||
                     currentIndex < map.getEnemiesInfo().at(currentInfo).index
                     ) {
-                    output << static_cast <const unsigned int>
-                        (Map::BlockTypes::None);
+                    output << static_cast <unsigned int> (Map::BlockTypes::None);
                 } else {
                     const JBTypes::Dir& dir =
                         map.getEnemiesInfo().at(currentInfo).enemy->direction();
@@ -1027,8 +1023,7 @@ void MapGenerator::verificationMap (std::ifstream& input, const Map& map) {
                     currentInfo >= map.getSpecialInfo().size() ||
                     currentIndex < map.getSpecialInfo().at(currentInfo).index
                     ) {
-                    output << static_cast <const unsigned int>
-                        (Map::BlockTypes::None);
+                    output << static_cast <unsigned int> (Map::BlockTypes::None);
                 } else {
                     const JBTypes::Dir& dir =
                         map.getSpecialInfo().at(currentInfo).special->direction();
